HORRIFIC CODE
Probably all the docs are wrong, error handling is non-existent,
nothing is tested. All this does is establish *a* solution, to be
fixed. Best to revert and re-add things in steps with proper tests
and docs.

Demo time!

In `flake.nix` I've added an example deployment (need to move that btw).
The `providers.local` definition should be moved to the
`nixops4-resources-local` flake.

The interesting part, the fundamental bits that expression authors get
to build on is the deployment function.

Imagine that besides `resources`, we have can declare module system
options that control the deployment configuration. The relevant part:

```nix
{
  nixops4Deployments.default = {
    # _type will be set by, say, a nixops4.mkDeployment function. TBD.
    _type = "nixops4Deployment";
    deploymentFunction = { resources }:
      let
        providers = ...;
      in
      {
        resources = {
          thefile = {
            provider = providers.local;
            type = "file";
            inputs = {
              name = "/tmp/hello-nixops4";
              contents = resources.hello.stdout;
              #          ^^^^^^^^^^^^^^^^^^^^^^
              # NixOps4 applies the `hello` resource and we get to use its value
              # basically anywhere, forming a directed acyclic graph of resource
              # information flows.
              # NixOps2 could not do this and had to rely on hacks like
              # placeholders that were hard to work with, and simply less capable
              # than true Nix values.
            };
          };
          hello = {
            provider = providers.local;
            type = "exec";
            inputs = {
              command = "bash";
              args = [ "-c" "echo 'Hello, NixOps4!'" ];
              stdin = null;
            };
          };
        };
      };
  };
}
```

And here's applying it:

```console
$ nixops4 apply
The following resources will be checked, created and/or updated:
  - thefile
  - hello

[... bunch of debug messages ...]

Done!
The following resources were created:
Resource hello:
  - input args: Array [String("-c"), String("echo 'Hello, NixOps4!'")]
  - input command: String("bash")
  - input stdin: Null
  - output stdout: String("Hello, NixOps4!\n")
Resource thefile:
  - input name: String("/tmp/hello-nixops4")
  - input contents: String("Hello, NixOps4!\n")

```

It will only get better from here.