feat: Add panic hook to handle it correctly
Handling panics is tricky business, but for a TUI app it is necessary,
or the messages would be unreadable at best; invisible in this case.
This commit tries to minimize the complexity of the handler.
Crucially the hook must stop the TUI thread, which puts a hard minimum on
the complexity and risk that our handler is exposed to.