maint: Ask GCC for system headers
In the previous update commit, the Nix `cc.version` didn't match the
in-output directory name anymore, causing a build failure.

By asking GCC properly, we are robust against such discrepancies.

The script isn't great, relying on GCC "log" output specifics, but it
seems that the GCC output has been like this for a while, and the
lines before and after seem to be intentional about their purpose...