feat: New async scheduling with more concurrency
This is a significant change to the eval_client code, and a complete
overhaul of the apply scheduling logic.

What this does is make the scheduling logic more flexible by letting
Bob handle most of the control flow for us.

As immediate benefits, the resource provider execution is now
concurrent, and we get cycle detection for (mutually) recursive
resource property dependencies.

(I should probably rename Bob)