refactor: schemafy -> typify
schemafy had an annoying bug related to naming, and it wasn't as
clever when it comes to recognizing and exploiting the pattern of
oneOfs as they're put into an enum.
The typify generated code is simpler and easier to use, and the
project has active maintainers.

I have mixed feelings about `build.rs`, but that's a solved problem
now, and we get to inspect the generated code a bit more easily,
at least with the current tooling.